Bootloader
Refactor whole project, splitting functions into bootloader and app.
The bootloader has a start screen with the list of installed apps, and it loads
apps by copying their code from flash to internal memory.
The bootloader has control of the execution, invoking the app through callbacks.
It also takes care of showing the low battery screen.
There are also several symbols in the bootloader that the app can link to.
Not tested on hardware yet.

Other changes:
- Complete refactor of gcprog to allow app actions: install, uninstall, list, etc
  Most previously existing commands were removed.
- Partial refactor of assets packer to allow simple extension and data in internal memory.
- graphics_image functions by image type instead of #ifdefs (image tests not updated yet).