TWorld: teleporter performance improvements
Cache teleporter positions to avoid scanning
the whole map each time an actor teleports.
This was especially problematic in CCLP2/68,
where after triggering a trap the gameplay went
to about 4 FPS. With this change, it's around 8.

For levels with >128 teleporters, the old method
is still used. For levels with no teleporters,
the old method is used to avoid scanning the whole
map for no reason.